+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Upcoming...
 
 - ... <!-- Add new lines here. -->
+- feat: Turn on `reportUnusedDisableDirectives` option (auto-fixable)
 - feat: Raise `@typescript-eslint/no-explicit-any` back to `error` level
 - fix: `prefer-const` switches to "error" in TS files (+ allow destructuring)
 - fix: Relax `react/jsx-no-useless-fragment` with `allowExpressions` option